pyproject.toml
[project]
name = "ve-edu-vault"
version = "0.1.1"
description = "Virtu Educate Backend to augment Edx"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
add-ons = [
    "adrf>=0.1.9",
    "bleach>=6.2.0",
    "cryptography>=45.0.2",
    "django-cors-headers>=4.6.0",
    "django-csp>=4.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-json-widget>=2.0.1",
    "django-oauth-toolkit>=3.0.1",
    "django-ratelimit>=4.1.0",
    "django-rest-knox",
    "pytz>=2025.2",
    "redis>=5.2.1",
]
ai = [
    "langchain>=0.3.17",
]
core = [
    "redis>=5.2.1",
    "celery>=5.4.0",
    "django>=5.2",
    "djangorestframework>=3.15.2",
    "uvicorn>=0.34.2",
    "pylti1p3>=2.0.0",
    "qstash>=2.0.5",
    "whitenoise>=6.9.0",
    "pandas>=2.3.1",
    "django-redis>=6.0.0",
]
database = [
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.13.0",
]
lint = [
    "black>=24.10.0",
    "flake8>=7.1.1",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
]
network = [
    "authlib>=1.6.0",
    "httpx>=0.28.1",
    "requests>=2.32.3",
    "requests-oauthlib>=2.0.0",
]
testing = [
    "factory-boy>=3.3.1",
    "faker>=33.1.0",
    "freezegun>=1.5.5",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.1",
    "pytest-sugar>=1.0.0",
    "pytest-xdist>=3.8.0",
]
utility = [
    "django-admin-env-notice>=1.0.1",
    "django-model-utils>=5.0.0",
    "ipython>=9.2.0",
    "jsonfield>=3.2.0",
    "pandas-stubs>=2.3.0.250703",
    "pydantic>=2.10.2",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.1",
    "swapper>=1.4.0",
]

[tool.uv.sources]
django-rest-knox = { git = "https://github.com/Virtu-E/django-rest-knox", rev = "5.0.3" }
